shift authinfo functions to users.ts
